requirements.txt
django
djangorestframework
djangorestframework-simplejwt
gunicorn